LPD-24115 Remove AOP proxy for CounterLocalService
