+++ portal-impl/src/com/liferay/counter/service/impl/CounterLocalServiceImpl.java
@@ -7,9 +7,6 @@ package com.liferay.counter.service.impl;
 
 import com.liferay.counter.kernel.service.CounterLocalService;
 import com.liferay.counter.service.base.CounterLocalServiceBaseImpl;
-import com.liferay.portal.kernel.transaction.Isolation;
-import com.liferay.portal.kernel.transaction.Propagation;
-import com.liferay.portal.kernel.transaction.Transactional;
 
 import java.util.List;
 
@@ -31,49 +28,31 @@ public class CounterLocalServiceImpl
 	}
 
 	@Override
-	@Transactional(
-		isolation = Isolation.COUNTER, propagation = Propagation.REQUIRES_NEW
-	)
 	public long increment() {
 		return counterFinder.increment();
 	}
 
 	@Override
-	@Transactional(
-		isolation = Isolation.COUNTER, propagation = Propagation.REQUIRES_NEW
-	)
 	public long increment(String name) {
 		return counterFinder.increment(name);
 	}
 
 	@Override
-	@Transactional(
-		isolation = Isolation.COUNTER, propagation = Propagation.REQUIRES_NEW
-	)
 	public long increment(String name, int size) {
 		return counterFinder.increment(name, size);
 	}
 
 	@Override
-	@Transactional(
-		isolation = Isolation.COUNTER, propagation = Propagation.REQUIRES_NEW
-	)
 	public void rename(String oldName, String newName) {
 		counterFinder.rename(oldName, newName);
 	}
 
 	@Override
-	@Transactional(
-		isolation = Isolation.COUNTER, propagation = Propagation.REQUIRES_NEW
-	)
 	public void reset(String name) {
 		counterFinder.reset(name);
 	}
 
 	@Override
-	@Transactional(
-		isolation = Isolation.COUNTER, propagation = Propagation.REQUIRES_NEW
-	)
 	public void reset(String name, long size) {
 		counterFinder.reset(name, size);
 	}
