+++ portal-impl/src/portal.properties
@@ -1461,17 +1461,6 @@
 ##
 
     #
-    # Set the transaction isolation level for the Counter service. See
-    # "transaction.isolation.portal" for more information. The preferred value
-    # is 8 when the transaction manager supports it. If the transaction manager
-    # does not support it, set this to -1 to use the database's default
-    # isolation level.
-    #
-    # Env: LIFERAY_TRANSACTION_PERIOD_ISOLATION_PERIOD_COUNTER
-    #
-    transaction.isolation.counter=2
-
-    #
     # Set the definition of the "portal" transaction isolation level. The
     # "portal" transaction isolation level is not a real isolation level. It is
     # just a pointer to a real isolation level that can be configured by setting
