+++ portal-impl/src/system.properties
@@ -432,7 +432,6 @@
         com.liferay.portal.dao.orm.hibernate.SessionFactoryImpl,\
         com.liferay.portal.kernel.dao.orm.SessionFactory,\
         com.liferay.portal.service.persistence.impl.*,\
-        com.liferay.portal.spring.transaction.CounterTransactionExecutor,\
         com.liferay.portal.verify.model.*,\
         com.sun.*,\
         java.*,\
