+++ portal-impl/test/unit/com/liferay/portal/spring/transaction/DefaultTransactionExecutorTest.java
@@ -5,13 +5,18 @@
 
 package com.liferay.portal.spring.transaction;
 
-import com.liferay.portal.kernel.transaction.TransactionAttribute;
+import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.portal.kernel.transaction.TransactionLifecycleListener;
 import com.liferay.portal.kernel.transaction.TransactionLifecycleManager;
-import com.liferay.portal.kernel.transaction.TransactionStatus;
+import com.liferay.portal.kernel.util.ProxyUtil;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Method;
+
 import java.util.Arrays;
+import java.util.Objects;
+import java.util.function.Predicate;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -21,12 +26,13 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.interceptor.TransactionAttribute;
 
 /**
  * @author Shuyang Zhou
  */
-public class DefaultTransactionExecutorTest
-	extends CounterTransactionExecutorTest {
+public class DefaultTransactionExecutorTest {
 
 	@ClassRule
 	@Rule
@@ -45,36 +51,132 @@ public class DefaultTransactionExecutorTest
 			_recordTransactionLifecycleListener);
 	}
 
-	@Override
 	@Test
 	public void testCommit() throws Throwable {
-		super.testCommit();
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager();
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(t -> false);
+
+		transactionExecutor.execute(transactionAttributeAdapter, () -> null);
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter,
+			RecordPlatformTransactionManager.TRANSACTION_STATUS, null);
 
 		_recordTransactionLifecycleListener.verify(null);
 	}
 
-	@Override
 	@Test
 	public void testCommitWithAppException() throws Throwable {
-		super.testCommitWithAppException();
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager();
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(throwable -> false);
+
+		try {
+			transactionExecutor.execute(
+				transactionAttributeAdapter,
+				() -> {
+					throw appException;
+				});
+
+			Assert.fail();
+		}
+		catch (Throwable throwable) {
+			Assert.assertSame(appException, throwable);
+		}
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter,
+			RecordPlatformTransactionManager.TRANSACTION_STATUS, null);
 
 		_recordTransactionLifecycleListener.verify(null);
 	}
 
-	@Override
 	@Test
 	public void testCommitWithAppExceptionWithCommitException()
 		throws Throwable {
 
-		super.testCommitWithAppExceptionWithCommitException();
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager() {
+
+				@Override
+				public void commit(TransactionStatus transactionStatus) {
+					ReflectionUtil.throwException(commitException);
+				}
+
+			};
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(throwable -> false);
+
+		try {
+			transactionExecutor.execute(
+				transactionAttributeAdapter,
+				() -> {
+					throw appException;
+				});
+
+			Assert.fail();
+		}
+		catch (Throwable throwable) {
+			Assert.assertSame(commitException, throwable);
+
+			Throwable[] throwables = commitException.getSuppressed();
+
+			Assert.assertEquals(
+				Arrays.toString(throwables), 1, throwables.length);
+			Assert.assertEquals(appException, throwables[0]);
+		}
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter, null, null);
 
 		_recordTransactionLifecycleListener.verify(commitException);
 	}
 
-	@Override
 	@Test
 	public void testCommitWithCommitException() throws Throwable {
-		super.testCommitWithCommitException();
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager() {
+
+				@Override
+				public void commit(TransactionStatus transactionStatus) {
+					ReflectionUtil.throwException(commitException);
+				}
+
+			};
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(throwable -> false);
+
+		try {
+			transactionExecutor.execute(
+				transactionAttributeAdapter, () -> null);
+
+			Assert.fail();
+		}
+		catch (Throwable throwable) {
+			Assert.assertSame(commitException, throwable);
+		}
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter, null, null);
 
 		_recordTransactionLifecycleListener.verify(commitException);
 	}
@@ -190,25 +292,152 @@ public class DefaultTransactionExecutorTest
 		}
 	}
 
-	@Override
+	@Test
+	public void testGetPlatformTransactionManager() {
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager();
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		Assert.assertSame(
+			recordPlatformTransactionManager,
+			transactionExecutor.getPlatformTransactionManager());
+	}
+
 	@Test
 	public void testRollbackOnAppException() throws Throwable {
-		super.testRollbackOnAppException();
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager();
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(
+				throwable -> throwable == appException);
+
+		try {
+			transactionExecutor.execute(
+				transactionAttributeAdapter,
+				() -> {
+					throw appException;
+				});
+
+			Assert.fail();
+		}
+		catch (Throwable throwable) {
+			Assert.assertSame(appException, throwable);
+		}
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter, null,
+			RecordPlatformTransactionManager.TRANSACTION_STATUS);
 
 		_recordTransactionLifecycleListener.verify(appException);
 	}
 
-	@Override
 	@Test
 	public void testRollbackOnAppExceptionWithRollbackException()
 		throws Throwable {
 
-		super.testRollbackOnAppExceptionWithRollbackException();
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager() {
+
+				@Override
+				public void rollback(TransactionStatus transactionStatus) {
+					ReflectionUtil.throwException(rollbackException);
+				}
+
+			};
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(
+				throwable -> throwable == appException);
+
+		try {
+			transactionExecutor.execute(
+				transactionAttributeAdapter,
+				() -> {
+					throw appException;
+				});
+
+			Assert.fail();
+		}
+		catch (Throwable throwable) {
+			Assert.assertSame(rollbackException, throwable);
+
+			Throwable[] throwables = rollbackException.getSuppressed();
+
+			Assert.assertEquals(
+				Arrays.toString(throwables), 1, throwables.length);
+			Assert.assertEquals(appException, throwables[0]);
+		}
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter, null, null);
 
 		_recordTransactionLifecycleListener.verify(appException);
 	}
 
-	@Override
+	@Test
+	public void testTransactionExecutorMethods() throws Throwable {
+		RecordPlatformTransactionManager recordPlatformTransactionManager =
+			new RecordPlatformTransactionManager();
+
+		TransactionExecutor transactionExecutor = createTransactionExecutor(
+			recordPlatformTransactionManager);
+
+		TransactionAttributeAdapter transactionAttributeAdapter =
+			_newTransactionAttributeAdapter(t -> t == appException);
+
+		assertTransactionExecutorThreadLocal(transactionExecutor, false);
+
+		TransactionStatusAdapter transactionStatusAdapter =
+			transactionExecutor.start(transactionAttributeAdapter);
+
+		assertTransactionExecutorThreadLocal(transactionExecutor, true);
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter, null, null);
+
+		try {
+			transactionExecutor.rollback(
+				appException, transactionAttributeAdapter,
+				transactionStatusAdapter);
+
+			Assert.fail();
+		}
+		catch (Exception exception) {
+			Assert.assertSame(appException, exception);
+		}
+
+		assertTransactionExecutorThreadLocal(transactionExecutor, false);
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter, null,
+			RecordPlatformTransactionManager.TRANSACTION_STATUS);
+
+		recordPlatformTransactionManager.setRollbackTransactionStatus(null);
+
+		transactionStatusAdapter = transactionExecutor.start(
+			transactionAttributeAdapter);
+
+		assertTransactionExecutorThreadLocal(transactionExecutor, true);
+
+		transactionExecutor.commit(
+			transactionAttributeAdapter, transactionStatusAdapter);
+
+		assertTransactionExecutorThreadLocal(transactionExecutor, false);
+
+		recordPlatformTransactionManager.verify(
+			transactionAttributeAdapter,
+			RecordPlatformTransactionManager.TRANSACTION_STATUS, null);
+	}
+
 	protected void assertTransactionExecutorThreadLocal(
 		TransactionExecutor transactionExecutor, boolean inTransaction) {
 
@@ -223,11 +452,38 @@ public class DefaultTransactionExecutorTest
 		}
 	}
 
-	@Override
 	protected TransactionExecutor createTransactionExecutor(
 		PlatformTransactionManager platformTransactionManager) {
 
 		return new DefaultTransactionExecutor(platformTransactionManager);
+	}
+
+	protected final Exception appException = new Exception();
+	protected final Exception commitException = new Exception();
+	protected final Exception rollbackException = new Exception();
+
+	private TransactionAttributeAdapter _newTransactionAttributeAdapter(
+		Predicate<Throwable> predicate) {
+
+		return new TransactionAttributeAdapter(
+			(TransactionAttribute)ProxyUtil.newProxyInstance(
+				TransactionAttribute.class.getClassLoader(),
+				new Class<?>[] {TransactionAttribute.class},
+				new InvocationHandler() {
+
+					@Override
+					public Object invoke(
+						Object proxy, Method method, Object[] args) {
+
+						if (Objects.equals(method.getName(), "rollbackOn")) {
+							return predicate.test((Throwable)args[0]);
+						}
+
+						throw new UnsupportedOperationException(
+							method.toString());
+					}
+
+				}));
 	}
 
 	private final RecordTransactionLifecycleListener
@@ -239,24 +495,31 @@ public class DefaultTransactionExecutorTest
 
 		@Override
 		public void committed(
-			TransactionAttribute transactionAttribute,
-			TransactionStatus transactionStatus) {
+			com.liferay.portal.kernel.transaction.TransactionAttribute
+				transactionAttribute,
+			com.liferay.portal.kernel.transaction.TransactionStatus
+				transactionStatus) {
 
 			throw new RuntimeException("commitThrowable");
 		}
 
 		@Override
 		public void created(
-			TransactionAttribute transactionAttribute,
-			TransactionStatus transactionStatus) {
+			com.liferay.portal.kernel.transaction.TransactionAttribute
+				transactionAttribute,
+			com.liferay.portal.kernel.transaction.TransactionStatus
+				transactionStatus) {
 
 			throw new RuntimeException("createThrowable");
 		}
 
 		@Override
 		public void rollbacked(
-			TransactionAttribute transactionAttribute,
-			TransactionStatus transactionStatus, Throwable throwable) {
+			com.liferay.portal.kernel.transaction.TransactionAttribute
+				transactionAttribute,
+			com.liferay.portal.kernel.transaction.TransactionStatus
+				transactionStatus,
+			Throwable throwable) {
 
 			throw new RuntimeException("rollbackThrowable");
 		}
@@ -268,24 +531,31 @@ public class DefaultTransactionExecutorTest
 
 		@Override
 		public void committed(
-			TransactionAttribute transactionAttribute,
-			TransactionStatus transactionStatus) {
+			com.liferay.portal.kernel.transaction.TransactionAttribute
+				transactionAttribute,
+			com.liferay.portal.kernel.transaction.TransactionStatus
+				transactionStatus) {
 
 			_committed = true;
 		}
 
 		@Override
 		public void created(
-			TransactionAttribute transactionAttribute,
-			TransactionStatus transactionStatus) {
+			com.liferay.portal.kernel.transaction.TransactionAttribute
+				transactionAttribute,
+			com.liferay.portal.kernel.transaction.TransactionStatus
+				transactionStatus) {
 
 			_created = true;
 		}
 
 		@Override
 		public void rollbacked(
-			TransactionAttribute transactionAttribute,
-			TransactionStatus transactionStatus, Throwable throwable) {
+			com.liferay.portal.kernel.transaction.TransactionAttribute
+				transactionAttribute,
+			com.liferay.portal.kernel.transaction.TransactionStatus
+				transactionStatus,
+			Throwable throwable) {
 
 			_throwable = throwable;
 		}
